Added Graft.branch().
